% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_info.R
\name{acommentary_info}
\alias{acommentary_info}
\title{acommentary_info}
\usage{
acommentary_info(KEYS, matchIDs, localteam, visitorteam)
}
\arguments{
\item{KEYS}{A list containing options such as testing / prediction /
important variables and information. Also contains API information.}

\item{matchIDs}{A character vector of matchIDs that match
the matchEvents.}

\item{localteam}{An integer ID value representing the home team
as defined by the API.}

\item{visitorteam}{An integer ID value representing the away team
as defined by the API.}
}
\value{
Returns nothing, a redis hash map is set with the
 commentary information and IDs are stored as a redis set.
}
\description{
A function that takes a KEYS$COMP and matchID's, and
 determines general match statistics for both local team and visitor
 team
}
\details{
Redis keys used;
  \itemize{
    \item{\strong{[SET]} :: \code{c_eventInSet:{comp_id}}}
    \item{\strong{[HASH]} :: \code{cme:{comp_id}:{match_id}:{event_id}}}
  }
}
