% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_info.R
\name{acomp_info}
\alias{acomp_info}
\title{acomp_info}
\usage{
acomp_info(KEYS)
}
\arguments{
\item{KEYS}{A list containing options such as testing / prediction
options and also the API query information such as url.}
}
\value{
returns competitionIDs, a redis hash is set with season IDs, and a
 redis set is created to store the current seasonIDs.
}
\description{
A function that checks a unique competition has been
 obtained and is added to a hashMap of the form ....
 -> competitions:set
}
\details{
API endpoints;
  \itemize{
    \item{\emph{"/competitions?Authorization={auth_id}"}}
  }

 Redis Keys used;
  \itemize{
    \item{\strong{[SET]} :: \code{competition:set}}
  }
}
