% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_info.R
\name{acomp_standings}
\alias{acomp_standings}
\title{acomp_standings}
\usage{
acomp_standings(KEYS)
}
\arguments{
\item{KEYS}{A list containing options such as testing / prediction
options and also the API query information such as url.}
}
\value{
Returns nothing, a redis hash map is set with the competition
 standing information.
}
\description{
A function that takes a KEYS$COMP and returns the current
 table information.
}
\details{
API endpoints;
  \itemize{
    \item{\emph{"/standings/{comp_id}?Authorization={auth_id}"}}
  }

 Redis Keys used;
  \itemize{
    \item{\strong{[HASH]} :: \code{comp:season:_standing_:{comp_id}:{season}}}
  }
}
