% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_info.R
\name{aevent_info}
\alias{aevent_info}
\title{aevent_info}
\usage{
aevent_info(KEYS, matchIDs, matchEvents)
}
\arguments{
\item{KEYS}{A list containing options such as testing / prediction
options and also the API query information such as url.}

\item{matchIDs}{An integer character vector of matchIDs that match
the matchEvents.}

\item{matchEvents}{A list of data frames containing individual events
grouped per match.}
}
\value{
Returns nothing, a redis hash map is set with the event information
 and IDs are stored as a redis set.
}
\description{
A function that takes a KEYS$COMP, matchID's, and
 a data frame containing match event information to be split up and
 added to redis as single events.
}
\details{
Redis Keys used;
  \itemize{
    \item{\strong{[SET]} :: \code{c_eventInSet:{comp_id}}}
    \item{\strong{[HASH]} :: \code{cme:{comp_id}:{match_id}:{event_id}}}
  }
}
