% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_for_deployment.R
\name{analyse_and_predict}
\alias{analyse_and_predict}
\title{Analyse Match and Event and Predict}
\usage{
analyse_and_predict(deployed = FALSE, addAll = TRUE)
}
\arguments{
\item{deployed}{A boolean value to indicate whether the
function is being run from a deployed environment.}
}
\value{
Nothing. Redis is updated.
}
\description{
A function that analyses matches and events only,
 then straight after makes match predictions based on the competition,
 this is to be run as a CRON job in deployment.
}
