% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_for_deployment.R
\name{analyse_data}
\alias{analyse_data}
\title{Analyse Match and Event}
\usage{
analyse_data(deployed = FALSE)
}
\description{
A function that analyses matches and events only,
 this is to be run as a CRON job in deployment.
}
