% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_for_deployment.R
\name{analyse_players}
\alias{analyse_players}
\title{Analyse Players}
\usage{
analyse_players(deployed = FALSE)
}
\description{
A function that analyses players only,
 this is to be run as a CRON job in deployment.
}
\details{
Redis Keys used;
  \itemize{
    \item{\strong{[LIST]} :: \code{analysePlayers}}
  }
}
