% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_info.R
\name{aplayer_info}
\alias{aplayer_info}
\title{aplayer_info}
\usage{
aplayer_info(KEYS, playerLength)
}
\arguments{
\item{KEYS}{A list containing options such as testing / prediction /
important variables and information. Also contains API information.}

\item{playerLength}{An integer value that defines the number of players to
analyse for a given list of ID's previously generated.}
}
\value{
Returns nothing. Redis is updated with player information
}
\description{
A function that takes a KEYS$COMP and length of players to
 analyse. The playerID's are popped from a Redis list and queried. The player
 stats are then stored in appropriate redis keys as necessary.
}
\details{
API endpoints;
  \itemize{
    \item{\emph{"/player/{player_id}?Authorization={auth_id}"}}
  }

 Redis Keys used;
  \itemize{
    \item{\strong{[LIST]} :: \code{analysePlayers}}
    \item{\strong{[HASH]} :: \code{ctps_**:{comp_id}:{team_id}:{player_id}:{season}}}
  }
}
