% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_info.R
\name{ateam_info}
\alias{ateam_info}
\title{ateam_info}
\usage{
ateam_info(competitionID, teamListLength, KEYS)
}
\arguments{
\item{competitionID}{An integer defining the competitionID that the
team belongs to.}

\item{teamListLength}{An integer value that defines how long the list
containing teamID's is TeamID's are then popped from the list as they
are anaylsed.}

\item{KEYS}{A list containing options such as testing / prediction
options and also the API query information such as url.}
}
\value{
Returns nothing. A Redis hash map is set with the team
 information.
}
\description{
A function that takes a competitionID and integer value
 with details of the teamID list for analysis. Each team is
 queried by the API for relevant information and statistics are
 stored.
}
\details{
API endpoints;
  \itemize{
    \item{\emph{"/team/{team_id}?Authorization={auth_id}"}}
  }

 Redis Keys used;
  \itemize{
    \item{\strong{[SET]} :: \code{c_playerSetInfo}}
    \item{\strong{[LIST]} :: \code{analyseTeams}}
    \item{\strong{[LIST]} :: \code{analysePlayers}}
    \item{\strong{[HASH]} :: \code{ct_basics:{comp_id}:{team_id}}}
    \item{\strong{[HASH]} :: \code{ct_stats:{comp_id}:{team_id}}}
    \item{\strong{[HASH]} :: \code{ctp:{comp_id}:{team_id}:{player_id}}}
  }
}
