% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_calculate_svm.R
\name{calculate_svm}
\alias{calculate_svm}
\title{calculate_svm}
\usage{
calculate_svm(competitionID, seasonStarting, commentaryKeys, commentaryNames,
  matchData, totalData = data.frame(stringsAsFactors = FALSE))
}
\arguments{
\item{competitionID}{An integer value denoting the competition ID.}

\item{seasonStarting}{An integer denoting the start year of the season.}

\item{commentaryKeys}{A vector of character values that hold the
commentary statistic hash maps in redis.}

\item{matchData}{A dataframe containing all the match data.}

\item{totalData}{A null dataframe which initialises the return value.}
}
\value{
Returns a dataframe containing the column names of 'returnItems'
 plus a few other metrics.
}
\description{
A function that takes current statistical data and combines
 it into a dataframe to be passed later to an SVM classifier.
}
