% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_all.R
\name{classify_all}
\alias{classify_all}
\title{classify_all}
\usage{
classify_all(competitionID, competitionName, seasonStarting, returnItems,
  matchLimit = 150, testing = FALSE, printToSlack, KEYS)
}
\arguments{
\item{matchLimit}{}
}
\value{
A list containing the best SVM calculated in the first key, and
 the best factors to use in the second key.
}
\description{
A function that loops through all possibilities of the variables
 to be included in the classifier. That means the recent form and results
 will be better represented as the best classifier will be built on the most
 recent results.
}
\details{
1) classify_recreate_matchdata():
  2) calculateSVMData() -> classify_match_result() -> { classify_team_form() -> classify_match_result() }
     -> calculateAdditionalMetrics():
  3) getBinns():
  4) optimizeSVM() -> calculateBestSVMFit()
     -> { classify_generate_predictions() ->
                 classify_emoji() ->
                 { classify_homeaway_stats() -> classify_commentary_stats() } ->
                 { classify_form_from_match() -> classify_match_result() } ->
                 mapFormToInteger() }
  5) predictFutureMatches() -> formatDates() -> getGeneralData() ->  { classify_generate_predictions() -> ... }
}
