% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_method_selection.R
\name{classify_method_selection}
\alias{classify_method_selection}
\title{Classify Method Selection}
\usage{
classify_method_selection(KEYS, method = "", singleFixture)
}
\description{
A function to decide which model to use
 during the classification
}
