% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functionality.R
\name{commentary_frame}
\alias{commentary_frame}
\title{Commentary Frame}
\usage{
commentary_frame(orderedKeys, commentaryNames)
}
\arguments{
\item{orderedKeys}{A character vector of commentary key names
that are stored in redis in some order.}

\item{commentaryNames}{A character vector defining the field
names of the key names to be returned back to store in the
new data frame.}
}
\value{
A data frame containing all the commentary information
 with the number of columns equal to that of the \code{commentaryNames}.
}
\description{
A function that takes ordered keys and
 the names from the commentary and takes them from redis
 and creates a data frame with number of rows equal to
 the length of the keys, provided they all exist and have
 non null values.
}
