% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_utils.R
\name{commentary_from_redis}
\alias{commentary_from_redis}
\title{Commentary From Redis}
\usage{
commentary_from_redis(KEYS, keyName, returnItems)
}
\arguments{
\item{keyName}{A character string that defines the redis key
where the commentary is stored.}

\item{returnItems}{A character vector defining the names of the
fields in the commentary key in redis to be retrieved.}
}
\value{
A vector of integers that correspond to the \code{returnItem}
 values requested.
}
\description{
A function that retrieves the basic commentary
 information from redis.
}
