% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_utils.R
\name{commentary_stats}
\alias{commentary_stats}
\title{commentary_stats}
\usage{
commentary_stats(commentary, returnItems)
}
\arguments{
\item{commentary}{A character vector of redis keys that hold a teams match
commentary.}

\item{returnItems}{A vector of character values that hold the names of
fields to be returned for the commentary statistics.}
}
\value{
A average statistics for a particular team.
}
\description{
A function that takes the commentary values stored in redis and
 calculates an average value for the list of variables in the key for that team.
}
