% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_headtohead_data.R
\name{create_headtohead_data}
\alias{create_headtohead_data}
\title{Create Head to Head Data}
\usage{
create_headtohead_data(KEYS, saveData = TRUE)
}
\details{
Need to port forward redis connection to local to
 get all the commentary information.
}
