% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_log_dir}
\alias{create_log_dir}
\title{Create Log Directory}
\usage{
create_log_dir()
}
\value{
Nothing, a log directory is created.
}
\description{
A function that checks for the global enviornment
 variable \code{FS_DEPLOYLOC}, if it exists then it creates a
 log directory structure at \code{FS_DEPLOYLOC/logs/}.
}
