% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sparse.R
\name{create_sparse}
\alias{create_sparse}
\title{Create Sparse}
\usage{
create_sparse(my.data, boundLen)
}
\description{
A function used for creating a sparse matrix
 with \code{(ncol - 1)} features and \code{1} column of
 results, i.e. W / L / D.
}
