% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates_from_yaml.R
\name{dates_from_yaml}
\alias{dates_from_yaml}
\title{Get Dates}
\usage{
dates_from_yaml(KEYS, tNow = Sys.Date())
}
\description{
A function that checks whether or not the season
 is currently active or not.
}
