% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_calculate_dataset.R
\name{feat_commentaries}
\alias{feat_commentaries}
\title{Commentary Feature}
\usage{
feat_commentaries(KEYS, matchID, teamIDs, commentaryNames)
}
\arguments{
\item{KEYS}{A list containing options such as testing / prediction /
important variables and information. Also contains API information.}
}
\details{
Redis Keys used;
  \itemize{
    \item{\strong{[HASH]} :: \code{cmt:_commentary:{comp_id}:{match_id}:{team_id}}}
  }
}
