% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functionality.R
\name{form_from_matchdata}
\alias{form_from_matchdata}
\title{Form From MatchData}
\usage{
form_from_matchdata(teamID, matchData)
}
\arguments{
\item{teamID}{A character string that defines the current
teamID that is under investigation.}

\item{matchData}{A data frame that contains rows of single matches
that have been played between two teams.}
}
\value{
An integer value representing the form.
}
\description{
A function that takes a matchData data.frame
 and a teamID and calculates the form cast as an integer
 value.
}
