% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_dates}
\alias{format_dates}
\title{API Date Formatter}
\usage{
format_dates(standardDateFormat)
}
\arguments{
\item{standardDateFormat}{A \code{Sys.Date()} value.}
}
\value{
An API style date which is of the form dd.mm.yyyy
}
\description{
A function that takes a date as generated by R
 and produces a date format used by the API.
}
