% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_generate_predictions.R
\name{generate_predictions}
\alias{generate_predictions}
\title{generate_predictions}
\usage{
generate_predictions(KEYS, fixtureList)
}
\arguments{
\item{KEYS}{A list containing options such as testing / prediction /
important variables and information. Also contains API information.}

\item{fixtureList}{A data frame containing match localteam vs. visitorteam
information which CAN include actual results if testing == TRUE.}
}
\value{
The agregated value of `correct`.
}
\description{
Another layer which can handle both normal fixture prediction
 and also predicting a test data set with known results.
}
