% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_league_table.R
\name{get_weeks}
\alias{get_weeks}
\title{Get Weeks}
\usage{
get_weeks(x)
}
\arguments{
\item{x}{A character vector of \code{cwt_l} style keys}
}
\value{
An integer vector of weeks that are defined in the redis keys.
}
\description{
A function that is used by the data league table functionality
 in order to extract the week information from the keys
}
