% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_utils.R
\name{homeaway_stats}
\alias{homeaway_stats}
\title{classify_homeaway_stats}
\usage{
homeaway_stats(competitionID, singleFixture, seasonStarting, localVisitor,
  returnItems, matchFieldNames, testing)
}
\arguments{
\item{singleFixture}{A single row data frame containing match localteam vs.
visitorteam information which CAN include actual results if testing == TRUE.}

\item{localVisitor}{A character vector which contains one of
c('localteam_id', 'visitorteam_id').}

\item{testing}{A boolean value which decides whether to read the actual result
of the match and compare with the classifier.}
}
\value{
A list of the current stats in key 1 and form in key 2.
}
\description{
A function to set up a match performance of two teams which returns
 their current statistics and their form.
}
