% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_ml_grid.R
\name{initialize_ml_grid}
\alias{initialize_ml_grid}
\title{Initialize ML Grid}
\usage{
initialize_ml_grid()
}
\description{
A function for returning the grid to be used
 in optimizing the ML models.
}
