% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys_for_testing.R
\name{keys_for_testing}
\alias{keys_for_testing}
\title{Keys For Testing}
\usage{
keys_for_testing(dbnum = 3)
}
\value{
A list of important key values, accessible through KEYS$...
}
\description{
A function that sets up the \code{KEYS}
 variable used throughout the package, with default parameters
 that avoid API calls and printing to screen etc.
}
