% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{Run script to populate Redis.}
\usage{
main(printToSlack = TRUE)
}
\details{
-- Summary of data structures stored in Redis --

 1.1) Match information:
 -->  [csm]:{comp_id}:{season}:{match_id} - [HASH]
 1.2) Match exists?
 -->  [c_matchSetInfo]:{comp_id} - [SET]
 1.3) Team exists?
 -->  [c_teamSetInfo]:{comp_id} - [SET]
 1.4) Match commentary
 -->  [cmt_commentary]:{comp_id}:{match_id}:{team_id} - [HASH]
 1.5) Player statistics per match
 -->  [cmp]:{comp_id}:{match_id}:{player_id} - [HASH]

 2.1) Events already analysed?
 -->  [c_eventInSet]:{comp_id} - [SET]
 2.2) Single event information:
 -->  [cme]:{comp_id}:{match_id}:{event_id} - [HASH]

 3.1) Basic team information:
 -->  [ct_basic]:{comp_id}:{team_id} - [HASH]
 3.2) Team statistics:
 -->  [ct_stats]:{comp_id}:{team_id} - [HASH]
 3.3) Player information:
 -->  [ctp]:{comp_id}:{team_id}:{player_id} - [HASH]

 4.1) Player statistics:
 -->  [ctps_[x]]:{comp_id}:{team_id}:{player_id}:{season} - [HASH]
      -->  where x = { club, club_intl, cups, national}
}
