% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_form_info.R
\name{match_result}
\alias{match_result}
\title{Match Result}
\usage{
match_result(scoreCurrent, scoreOther)
}
\arguments{
\item{scoreCurrent}{An integer value denoting the home team score.}

\item{scoreOther}{An integer value denoting the away team score.}
}
\value{
Returns one of 'W' / 'L' / 'D'.
}
\description{
A function that returns a single character value of
 'W' / 'L' / 'D' depending on the scores and which team scored them.
}
