% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_network.R
\name{neural_network}
\alias{neural_network}
\title{Generate Artifical Neural Network}
\usage{
neural_network(totalData)
}
\arguments{
\item{totalData}{A data set that contains scaled data
and a vector of results as a column in the data frame.}
}
