% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_features.R
\name{optimize_features}
\alias{optimize_features}
\title{Optimize Features}
\usage{
optimize_features(data.set, team = 100, player = 0, days = 4, decay = 0,
  folds = 10)
}
\description{
A function used to optimize data set features.
}
\details{
This function takes a number of parameters, and goes through
 each unique combination in order to try and optimize the feature sets
 based on the total data used.
}
