% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_svm_info.R
\name{optimize_svm}
\alias{optimize_svm}
\title{optimize_svm}
\usage{
optimize_svm(competitionID, totalData, seasonStarting, testData, matchData,
  binList, returnItems, matchFieldNames, testing = TRUE)
}
\arguments{
\item{totalData}{A dataframe containing the data used by the classifier with
variables matching those of `binList`.}

\item{testData}{A dataframe containing the test set used to find the results
of the classification by comparing with the results in the testData.}

\item{binList}{A list of current interval values for the allowed variables used
for building the classifier.}

\item{returnItems}{A vector of character values that hold the names of
fields to be returned for the commentary statistics.}

\item{testing}{A boolean value which decides whether to read the actual result
of the match and compare with the classifier.}
}
\value{
A list containing the best SVM calculated in the first key, and
 the best factors to use in the second key.
}
\description{
A function that loops through all possibilities of the variables
 to be included in the classifier. That means the recent form and results
 will be better represented as the best classifier will be built on the most
 recent results.
}
