% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_variables.R
\name{optimize_variables}
\alias{optimize_variables}
\title{Optimize Variables}
\usage{
optimize_variables(total.metrics, DAYS = c(3, 4, 5), GRID_PTS = c(2, 4, 6,
  8, 10), GRID_BOUND = c(0.01, 0.05, 0.1, 0.15, 0.2), DECAY = c(0.5, 0.75,
  1, 1.25, 1.5, 2, 2.5), TOTAL_PERC = seq(from = 0, to = 1, by = 0.1),
  REP = 1, THRESH = 0.01)
}
