% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newfs.R
\name{ord_keys}
\alias{ord_keys}
\title{ord keys}
\usage{
ord_keys(commentaryKeys, KEYS)
}
\arguments{
\item{KEYS}{A list containing options such as testing / prediction /
important variables and information. Also contains API information.}
}
