% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_commentaries.R
\name{order_commentaries}
\alias{order_commentaries}
\title{Order Commentaries}
\usage{
order_commentaries(KEYS, commentaryKeys)
}
\arguments{
\item{KEYS}{A list containing options such as testing / prediction /
important variables and information. Also contains API information.}

\item{commentaryKeys}{A character vector of commentary key names as
found in redis.}
}
\value{
A character vector of the same commentaryKeys but ordered
 by date in which they occured.
}
\description{
A function that takes a vector of commentary key names
 that are stored in redis, the date is checked to see when they happened
 and then that vector is ordered and sorted and returned back in the
 correct order.
}
\details{
Redis Keys used;
  \itemize{
    \item{\strong{[HASH]} :: \code{csm:{comp_id}:{season}:{match_id}}}
  }
}
