% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_utils.R
\name{order_matchdata}
\alias{order_matchdata}
\title{Order Match Dataset}
\usage{
order_matchdata(matchData)
}
\arguments{
\item{matchData}{A data frame that contains rows of single matches
that have been played between two teams.}
}
\value{
A data frame that has been ordered by date.
}
\description{
A function that takes an arbitrary data frame
 consisting of match data and orders it by date in ascending
 order.
}
