% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme.R
\name{plot_theme}
\alias{plot_theme}
\title{Plot Theme}
\usage{
plot_theme(fontFamily = "Purisa", titleFont = 26)
}
\arguments{
\item{fontFamily}{A character string that defines the font
family to be used for the axis / title / legend etc.}
}
\description{
Holds the theme for the plots to be used in the
 README
}
