% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{possible_env}
\alias{possible_env}
\title{Possible Environment}
\usage{
possible_env()
}
\value{
A character string of R global environment location paths.
}
\description{
A function that searches typical R global
 environment locations, to try and read the values
 defining the API and other sensitive data.
}
