% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_for_deployment.R
\name{predict_fixtures}
\alias{predict_fixtures}
\title{Predict Fixtures}
\usage{
predict_fixtures(deployed = FALSE)
}
\arguments{
\item{deployed}{A boolean value to indicate whether the
function is being run from a deployed environment.}
}
\value{
Nothing. Redis is updated.
}
\description{
A function to be run as a CRON job during
 deployment. It gathers some data from redis, tries to
 project features and then builds a data set which is
 run against some predefined (or dynamic) model carried
 in the package.
}
