% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_predict_matches.R
\name{predict_matches}
\alias{predict_matches}
\title{classify_predict_matches}
\usage{
predict_matches(competitionID, competitionName, KEYS)
}
\arguments{
\item{competitionID}{An integer value denoting the competition ID.}
}
\value{
Nothing. Print results to screen.
}
\description{
A function that queries the endpoint and redis to obtain
 upcoming matches then with the current optimized SVM, attempts to predict
 the results.
}
