% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_vs_real.R
\name{predict_vs_real}
\alias{predict_vs_real}
\title{Predictions vs Actual Results}
\usage{
predict_vs_real(KEYS, readyToAnalyse, matches)
}
\arguments{
\item{KEYS}{A list containing options such as testing / prediction /
important variables and information. Also contains API information.}

\item{readyToAnalyse}{A character vector of the form of the redis keys.}

\item{matches}{A data frame that contains rows of single matches
that have been played between two teams.}
}
\value{
Nothing. The prediction key for a particular match is updated with
 a T or F whether the prediction was true or false.
}
\description{
A function that takes the prediction keys and the gathered
 matchData and decides whether the match has been played and then decides
 whether the match was a 'W' / 'L' / 'D' pivoted on the home team,
 and checks the prediction if it was correct (T) or incorrect (F).
}
\details{
Redis Keys used;
  \itemize{
    \item{\strong{[HASH]} :: \code{csdm_pred:{comp_id}:{season}:{month}:{match_id}}}
  }
}
