% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_projections.R
\name{project_form}
\alias{project_form}
\title{Commentary Projection}
\usage{
project_form(KEYS, teamIDs)
}
\arguments{
\item{KEYS}{A list containing options such as testing / prediction /
important variables and information. Also contains API information.}
}
