% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_information.R
\name{prs_comp}
\alias{prs_comp}
\title{Parse Competition}
\usage{
prs_comp(mComps)
}
\arguments{
\item{mComps}{A character vector of competitionIDs.}
}
\value{
A character vector of unique competitionIDs.
}
\description{
A function that parses competitionIDs
 and returns only the unique IDs.
}
