% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_information.R
\name{prs_season}
\alias{prs_season}
\title{Parse Season}
\usage{
prs_season(mSeasons)
}
\arguments{
\item{mSeasons}{A character vector of the form
2017/2018. (Can also take multiple, not sure of the
benefit just yet)}
}
\value{
A character string of yyyy format.
}
\description{
A function that takes a season form
 and returns the correct season starting value.
}
