% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_utils.R
\name{recreate_matchdata}
\alias{recreate_matchdata}
\title{Recreate Match Data}
\usage{
recreate_matchdata(competitionID, seasonStarting, matchLimit)
}
\arguments{
\item{competitionID}{An integer containing the competitionID that the
teams and match information belong to.}

\item{seasonStarting}{An integer defining the lower year for details on a season.}

\item{redisData}{An environment that defines the redis configuration where data is
to be searched for.}
}
\value{
matchData. A data frame containing all the matches in a particular season.
}
\description{
A function that provided with a competitionID and season start year,
 can auto generate and order by date all the basic match information
}
\details{
A search for all matches in a particular subset is made in Redis, a data frame
 is then constructed to rebuild the original API query and ordered by date.
}
