% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_form_info.R
\name{relative_form}
\alias{relative_form}
\title{Relative Form}
\usage{
relative_form(matchInfo, totalForm)
}
\arguments{
\item{matchInfo}{A list / data frame object that is a slice of data
of the form of matchData, which contains useful match information and
the date.}

\item{totalForm}{A named list object which contains the \code{$date} and \code{$form}
values, each vector is of equal length.}
}
\value{
An integer value of the current relative form as an integer.
}
\description{
A function that calculates the relative form from an
 arbitrary date, derived from the current date found in matchInfo and
 the data frame of total results and the dates which they happened.
}
