% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sensitive_keys}
\alias{sensitive_keys}
\title{sensitive_keys}
\usage{
sensitive_keys()
}
\arguments{
\item{None}{}
}
\value{
Returns nothing.
}
\description{
A function that loads all sensitive information into a
 global namespace for use throughout the code.
}
\details{
Set up:
 1) Fill in the appropriate sensitive information.
 2) Copy this file to sensitiveInfo.R and rename the
    function to sensitiveInfo()
}
