% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_from_yaml.R
\name{stats_from_yaml}
\alias{stats_from_yaml}
\title{}
\usage{
stats_from_yaml()
}
\description{
A function to insert statistics not returned
 by the API from a well-defined yaml file.
}
\details{
This function takes an input list as defined in
 \code{missing_data/statistics.yaml} and reads the home and away
 statistics, makes sure they exist in redis, and if they dont
 then it reads the values and populates by inferring the competition
 etc.
}
