% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_form_info.R
\name{team_form}
\alias{team_form}
\title{classify_team_form}
\usage{
team_form(matchData, teamID)
}
\arguments{
\item{matchData}{A dataframe containing all the match data, so it can
be subsetting easily during the vectorized function.}

\item{teamID}{A character string that denotes the current team being
analysed.}
}
\value{
Returns two lists nested inside one list.
}
\description{
A function that generates both the team form and date of
 the match in a list. A list is generated for both of these in terms
 of a home and away list.
}
