% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{teams_in_league}
\alias{teams_in_league}
\title{Teams in league}
\usage{
teams_in_league(compID)
}
\value{
A list of matching competition ids and team numbers
}
\description{
To normalise some of the data we need to know how
 many teams are present within each league.
}
