% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_league_table.R
\name{weekly_positions}
\alias{weekly_positions}
\title{Weekly Positions}
\usage{
weekly_positions(KEYS)
}
\arguments{
\item{KEYS}{A list containing options such as testing / prediction /
important variables and information. Also contains API information.}
}
\details{
Redis Keys used;
  \itemize{
    \item{\strong{[HASH]} :: \code{cwt_l:{comp_id}:{season}:{week#}:{team_id}}}
    \item{\strong{[HASH]} :: \code{cw_pl:{comp_id}:{season}:{week#}}}
  }
}
