% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_info.R
\name{acommentary_info}
\alias{acommentary_info}
\title{acommentary_info}
\usage{
acommentary_info(competitionID, matchIDs, localteam, visitorteam, KEYS,
  bypass = FALSE)
}
\arguments{
\item{competitionID}{An integer defining the competitionID that the
team belongs to.}

\item{matchIDs}{A character vector of matchIDs that match
the matchEvents.}
}
\value{
Returns nothing, a redis hash map is set with the
 commentary information and IDs are stored as a redis set.
}
\description{
A function that takes a competitionID and matchID's, and
 determines general match statistics for both local team and visitor
 team
}
\details{
EventID's are checked if they have been analysed already;
    ->   [c_eventInSet]:{comp_id}   ->   [SET]
 The actual event information is stored as a hash map as;
    ->   [cme]:{comp_id}:{match_id}:{event_id}  ->   [HASH]
}
