% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_info.R
\name{acomp_info}
\alias{acomp_info}
\title{acomp_info}
\usage{
acomp_info(KEYS, bypass = FALSE)
}
\arguments{
\item{production}{A boolean to indicate whether production (default)
runs are performed or testing carried out.}

\item{seasonIDs}{A list containing seasonIDs...}
}
\value{
returns competitionIDs, a redis hash is set with season IDs, and a
 redis set is created to store the current seasonIDs.
}
\description{
A function that checks a unique competition has been
 obtained and is added to a hashMap of the form ....
 -> competitions:set
}
