% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_info.R
\name{acomp_standings}
\alias{acomp_standings}
\title{acomp_standings}
\usage{
acomp_standings(competitionID, KEYS, bypass = FALSE)
}
\arguments{
\item{competitionID}{An integer containing the competition ID that the
teams and match information belong to.}
}
\value{
Returns nothing, a redis hash map is set with the competition
 standing information.
}
\description{
A function that takes a competitionID and returns the current
 table information.
}
\details{
Competition table information is stored in the following redis hash
   ->   comp:season:_standing_:{comp_id}:{season}
}
