% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_all.R
\name{add_all}
\alias{add_all}
\title{Add All Information}
\usage{
add_all(competitionID, seasonStarting, KEYS)
}
\arguments{
\item{competitionID}{An integer ID of the competition defined by the API.}

\item{seasonStarting}{An integer of _Y_ format that defines the start of the
season.}

\item{KEYS}{A list containing options such as testing / prediction
options and also the API query information such as url.}
}
\value{
Returns nothing.
}
\description{
A function that is called from a shell script to kick off
 the storing of important data and also any machine learning mechanisms
 for prediction.
}
\details{
This main function is split into 3 main important components
 \itemize{
   \item{
     All libraries, global variables are loaded. Storing mechanisms
     that incorporate new ID keys are generated to store all useful
     information for classification / algorithms.
   }
   \item{
     Statistical models are built on the stored data and redis, future
     fixtures are obtained and built into the models.
   }
   \item{
     Predictions are made based on the current data and models and
     anything useful is sent via slack for easy analysis.
   }
 }
}
