% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_all.R
\name{add_all}
\alias{add_all}
\title{add_all}
\usage{
add_all(competitionID, seasonStarting, KEYS)
}
\value{
Returns nothing.
}
\description{
A function that is called from a shell script to kick off
 the storing of important data and also any machine learning mechanisms
 for prediction.
}
\details{
This main function is split into 3 main important components
 1) All libraries, global variables are loaded. Storing mechanisms
    that incorporate new ID keys are generated to store all useful
    information for classification / algorithms.
 2) Statistical models are built on the stored data and redis, future
    fixtures are obtained and built into the models.
 3) Predictions are made based on the current data and models and
    anything useful is sent via slack for easy analysis.
 Example queries carried out:
    ->   matches?comp_id=1204&from_date=01.01.2016&to_date=20.06.2016
    ->   team/9092
}
