% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_info.R
\name{amatch_info}
\alias{amatch_info}
\title{amatch_info}
\usage{
amatch_info(KEYS)
}
\arguments{
\item{KEYS}{A list containing options such as testing / prediction /
important variables and information. Also contains API information.}
}
\value{
Returns a match dataframe containing all match information to update
 events in a particular match. Redis is updated with match information.

Returns a NULL dataframe if no matches are found.
}
\description{
A function that takes a KEYS$COMP and season year to query
 for all the matches in a particular season and saves new teams to a set for
 later analysis.
}
\details{
API endpoints;
  \itemize{
    \item{\emph{"/matches?comp_id={comp_id}&from_date={dd.mm.yyyy}&to_date={dd.mm.yyyy}&Authorization={auth_id}"}}
  }

 Redis Keys used;
  \itemize{
    \item{\strong{[LIST]} :: \code{analyseTeams}}
    \item{\strong{[SET]} :: \code{c_matchSetInfo:{comp_id}}}
    \item{\strong{[SET]} :: \code{c:{comp_id}:ready}}
    \item{\strong{[HASH]} :: \code{csm:{comp_id}:{season}:{match_id}}}
    \item{\strong{[HASH]} :: \code{c:{comp_id}:pred:{match_id}}}
  }
}
