% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_info.R
\name{aplayer_info}
\alias{aplayer_info}
\title{aplayer_info}
\usage{
aplayer_info(playerLength, currentSeasonYear, KEYS)
}
\arguments{
\item{playerLength}{An integer value that defines the number of players to
analyse for a given list of ID's previously generated.}
}
\value{
Returns nothing. Redis is updated with player information
}
\description{
A function that takes a competitionID and length of players to
 analyse. The playerID's are popped from a Redis list and queried. The player
 stats are then stored in appropriate redis keys as necessary.
}
\details{
Player stats infromation is stored in a hash map as;
    ->   [ctps]:{comp_id}:{team_id}:{player_id}:{season}:[_stats_`statType`_]   ->   [HASH]
}
