% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_info.R
\name{ateam_info}
\alias{ateam_info}
\title{ateam_info}
\usage{
ateam_info(competitionID, teamListLength, KEYS)
}
\arguments{
\item{competitionID}{An integer defining the competitionID that the
team belongs to.}

\item{teamListLength}{An integer value that defines how long the list
containing teamID's is TeamID's are then popped from the list as they
are anaylsed.}
}
\value{
Returns nothing. A Redis hash map is set with the team
 information.
}
\description{
A function that takes a competitionID and integer value
 with details of the teamID list for analysis. Each team is
 queried by the API for relevant information and statistics are
 stored.
}
\details{
A number of information is stored.
 The basic information is stored as;
    ->   [ct_basic]:{comp_id}:{team_id}   ->   [HASH]
 The team statistics is stored as;
    ->   [ct_stats]:{comp_id}:{team_id}   ->   [HASH]
 Player information relevent to the team is stored as;
    ->   [ctp]:{comp_id}:{team_id}:{player_id}   ->   [HASH]
}
