% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_calculate_dataset.R
\name{feat_form}
\alias{feat_form}
\title{Relative Form Feature}
\usage{
feat_form(matchData, teamIDs, singleMatchInfo)
}
\arguments{
\item{matchData}{A data frame...}

\item{teamIDs}{An integer vector that contains c(home_id, away_id).}

\item{singleMatchInfo}{A data frame with one row that contains
important match information between two teams.}
}
