% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_calculate_dataset.R
\name{feat_position}
\alias{feat_position}
\title{Relative Position Feature}
\usage{
feat_position(KEYS, matchID, teamIDs, matchDate = NULL)
}
\arguments{
\item{KEYS}{A list containing options such as testing / prediction /
important variables and information. Also contains API information.}
}
\details{
Redis Keys used;
  \itemize{
    \item{\strong{[KEY]} :: \code{c_startDate:{comp_id}:{season}}}
    \item{\strong{[HASH]} :: \code{csm:{comp_id}:{season}:{match_id}}}
  }
}
