% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newfs.R
\name{form_from_matchdata}
\alias{form_from_matchdata}
\title{Form From MatchData}
\usage{
form_from_matchdata(teamID, matchData)
}
\arguments{
\item{teamID}{...}

\item{matchData}{...}
}
\value{
An integer value
}
\description{
A function that takes a matchData data.frame
 and a teamID and calculates the form cast as an integer
 value.
}
