% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_form_info.R, R/optimize_calculation.R
\name{form_to_int}
\alias{form_to_int}
\alias{form_to_int}
\title{From Vector To Integer}
\usage{
form_to_int(oldForms, winPoints = 2, drawPoints = 1, losePoints = 0)

form_to_int(oldForms, winPoints = 2, drawPoints = 1, losePoints = 0)
}
\arguments{
\item{oldForms}{A character string which contains either `W`, `L`, or `D`
of arbitrary length}

\item{winPoints}{An integer defining the points accredited for a win.}

\item{drawPoints}{Same as above for a draw.}

\item{losePoints}{Same as above for a loss.}
}
\value{
An integer value defining the value of a teams form
}
\description{
A function that converts a form vector e.g. 'WLD' into an integer
 value defined by the function default values for win / lose / draw.
}
\examples{
 footballstats::form_to_int('WLDW')

}
