% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{get_data}
\usage{
get_data(endpoint, KEYS)
}
\arguments{
\item{endpoint}{A character string that defines the endpoint and parameters
utilized to gather the data. The endpoint is accessed by concatenating the
string HOST -> endpoint -> apiKey.}

\item{KEYS}{A list containing options such as testing / prediction /
important variables and information. Also contains API information.}
}
\value{
A list of data corresponding to the `endpoint` if status_code == 200

Null for an appropriate response if status_code != 200
}
\description{
A function that generates JSON data from a particular
 endpoint provided to the function as a string.
}
