% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_form_info.R
\name{match_result}
\alias{match_result}
\title{classify_match_result}
\usage{
match_result(scoreCurrent, scoreOther)
}
\arguments{
\item{scoreHome}{An integer value denoting the home team score.}

\item{scoreAway}{An integer value denoting the away team score.}

\item{homeOrAway}{A character vector which is either 'localteam_id'
or 'visitorteam_id', to conduct the correct operation on the two
score parameters.}
}
\value{
Returns one of 'W' / 'L' / 'D'.
}
\description{
A function that returns a single character value of
 'W' / 'L' / 'D' depending on the scores and which team scored them.
}
