% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_vs_real.R
\name{monthly_report}
\alias{monthly_report}
\title{Create Monthly Report}
\usage{
monthly_report(KEYS, month, year)
}
\arguments{
\item{KEYS}{A list containing options such as testing / prediction /
important variables and information. Also contains API information.}
}
\details{
This function is triggered via a CRON job,
 and prints the results from a months worth of data
 to the slack channel #reports
}
