% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_network.R
\name{neural_network}
\alias{neural_network}
\title{Generate Artifical Neural Network}
\usage{
neural_network(totalData, odds.results, FOLD_DATA, NN, LOGS = FALSE)
}
\arguments{
\item{totalData}{A data set that contains scaled data
and a vector of results as a column in the data frame.}
}
\value{
A neural network model built from \code{totalData}
 which is of class 'nn'.
}
\description{
A function that takes a scaled data set
 builds a neural network and reports on the accuracy
 of the built model.
}
