% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newfs.R
\name{order_commentaries}
\alias{order_commentaries}
\title{Order Commentaries}
\usage{
order_commentaries(KEYS, commentaryKeys)
}
\arguments{
\item{KEYS}{A list containing options such as testing / prediction /
important variables and information. Also contains API information.}
}
