% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_utils.R
\name{order_matchdata}
\alias{order_matchdata}
\title{Order Match Dataset}
\usage{
order_matchdata(matchData, formatter = "\%d.\%m.\%Y")
}
\arguments{
\item{matchData}{A data frame that contains rows of single matches
that have been played between two teams.}

\item{formatter}{A string that defines how the dates coming in should
look, i.e. allow for any kind of standard date formats.}
}
\value{
A data frame that has been ordered by date.
}
\description{
A function that takes an arbitrary data frame
 consisting of match data and orders it by date in ascending
 order.
}
