% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_predict_matches.R
\name{predict_matches}
\alias{predict_matches}
\title{Predict Matches}
\usage{
predict_matches(KEYS)
}
\arguments{
\item{KEYS}{A list containing options such as testing / prediction /
important variables and information. Also contains API information.}
}
\value{
Nothing. Print results to screen.
}
\description{
A function that queries the endpoint and redis to obtain
 upcoming matches then with the current optimized SVM, attempts to predict
 the results.
}
\details{
API endpoints;
  \itemize{
    \item{\emph{"/matches?comp_id={comp_id}&from_date={mm.dd.yyyy}&to_date={mm.dd.yyyy}&Authorization={auth_id}"}}
  }
}
