% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_vs_real.R
\name{predict_vs_real}
\alias{predict_vs_real}
\title{Predictions vs Actual Results}
\usage{
predict_vs_real(KEYS, readyToAnalyse, matches)
}
\arguments{
\item{KEYS}{A list containing options such as testing / prediction /
important variables and information. Also contains API information.}
}
\details{
Redis Keys used;
  \itemize{
    \item{\strong{[HASH]} :: \code{csdm_pred:{comp_id}:{season}:{month}:{match_id}}}
  }
}
