% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_projections.R
\name{project_commentaries}
\alias{project_commentaries}
\title{Commentary Projection}
\usage{
project_commentaries(KEYS, teamIDs, matchDate, matchID)
}
\arguments{
\item{KEYS}{A list containing options such as testing / prediction /
important variables and information. Also contains API information.}
}
\details{
Redis Keys used;
  \itemize{
    \item{\strong{[HASH]} :: \code{cmt_commentary:{comp_id}:{match_id}:{team_id}}}
  }
}
