% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_projections.R
\name{project_form}
\alias{project_form}
\title{Form Projection}
\usage{
project_form(KEYS, teamIDs)
}
\arguments{
\item{KEYS}{A list containing options such as testing / prediction /
important variables and information. Also contains API information.}

\item{teamIDs}{A character vector of length two, containing the home team
and away team in that order.}
}
\value{
A data frame that contains two columns, `form.h` and `form.a`
}
\description{
A function that projects the form of the two teams to the
 build up of this match.
}
\details{
Redis Keys used;
  \itemize{
    \item{\strong{[HASH]} :: \code{cmt_commentary:{comp_id}:{match_id}:{team_id}}}
    \item{\strong{[HASH]} :: \code{csm:{comp_id}:{season}:{match_id}}}
  }
}
