% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_information.R
\name{prs_ftscore}
\alias{prs_ftscore}
\title{Parse Full Time score}
\usage{
prs_ftscore(ftScore)
}
\arguments{
\item{ftScore}{A character string of the form
'[2-0]', and will be parsed as an integer vector of
2 0.}
}
\value{
An integer vector of the form c(home_score, away_score).
}
\description{
A function that takes a full time result and
 tries to parse this as an integer vector of H vs. A.
}
