% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_utils.R
\name{recreate_matchdata}
\alias{recreate_matchdata}
\title{Recreate Match Data}
\usage{
recreate_matchdata(KEYS, matchLimit = 10000)
}
\arguments{
\item{KEYS}{A list containing options such as testing / prediction /
important variables and information. Also contains API information.}
}
\value{
matchData. A data frame containing all the matches in a particular season.
}
\description{
A function that provided with a competitionID and season start year,
 can auto generate and order by date all the basic match information
}
