% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_utils.R
\name{recreate_matchdata}
\alias{recreate_matchdata}
\title{Recreate Match Data}
\usage{
recreate_matchdata(KEYS)
}
\arguments{
\item{KEYS}{A list containing options such as testing / prediction /
important variables and information. Also contains API information.}
}
\value{
matchData. A data frame containing all the matches in a particular season.
}
\description{
A function that can auto generate and order by date all
 the basic match information if it exists in redis.
}
\details{
Redis Keys used;
  \itemize{
    \item{\strong{[HASH]} :: \code{csm:{comp_id}:{season}:{match_id}}}
  }
}
