% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sensitive_keys}
\alias{sensitive_keys}
\title{Sensitive Keys}
\usage{
sensitive_keys(printToSlack, printToScreen, testing, storePred)
}
\arguments{
\item{printToSlack}{A boolean on whether results should be returned
to slack or not.}

\item{printToScreen}{A boolean on whether results should be printed
to screen or sink file or not.}

\item{testing}{A boolean to indicate whether tests are being run,
so that preloaded data sets are used and API endpoints are generally
avoided to test functionality rather than the endpoints.}

\item{storePred}{A boolean to indicate whether to store predicted
results in the prediction hash in redis.}
}
\value{
Returns a list of KEYS used by most functionality of \code{footballstats}.
}
\description{
A function that loads all sensitive information into a
 global namespace for use throughout the code.
}
