% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_form_info.R
\name{team_form}
\alias{team_form}
\title{Team Form}
\usage{
team_form(matchData, teamID)
}
\arguments{
\item{matchData}{A data frame that contains rows of single matches
that have been played between two teams.}

\item{teamID}{A character string that denotes the current team being
analysed.}
}
\value{
Returns a named list with `date` and `form`.
}
\description{
A function that generates both the team form and date of
 a particular team.
}
